move Link tests to separate file
