+++ packages/zent/__tests__/button.js
@@ -4,7 +4,7 @@ import {
   findRenderedDOMComponentWithTag,
   Simulate,
 } from 'react-dom/test-utils';
-import { Button, ButtonGroup, ButtonDirective, Link } from 'button';
+import { Button, ButtonGroup, ButtonDirective } from 'button';
 import Icon from 'icon';
 
 class Wrapper extends React.Component {
@@ -314,37 +314,5 @@ describe('<ButtonGroup />', () => {
     );
     const node = findRenderedDOMComponentWithTag(tree, 'div');
     expect(node.style.fontSize).toBe('20px');
-  });
-
-  test('Link', () => {
-    const url = 'https://www.youzan.com/404';
-    const target = '_blank';
-    const tree = renderIntoDocument(
-      <Wrapper>
-        <Link href={url} target={target} className="my-link">
-          link
-        </Link>
-      </Wrapper>
-    );
-    const node = findRenderedDOMComponentWithTag(tree, 'a');
-    expect(node.href).toBe(url);
-    expect(node.target).toBe(target);
-    expect(node.classList.contains('my-link')).toBe(true);
-  });
-
-  test('Link disabled', () => {
-    const tree = renderIntoDocument(
-      <Wrapper>
-        <Link href="http://www.youzan.com" disabled>
-          link
-        </Link>
-      </Wrapper>
-    );
-    const node = findRenderedDOMComponentWithTag(tree, 'a');
-    const preventDefault = jest.fn();
-    const stopPropagation = jest.fn();
-    Simulate.click(node, { preventDefault, stopPropagation });
-    expect(stopPropagation.mock.calls.length).toBe(1);
-    expect(preventDefault.mock.calls.length).toBe(1);
   });
 });
